updateValidationStateForBegruendung in differenceDialogUtils ausgelagert
