+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/common/differenceDialogUtils.ts
@@ -1,13 +1,19 @@
+import type { DifferenceBegruendung } from "@/types/ergebnismeldung/common/DifferenceBegruendung.ts";
+
 import { storeToRefs } from "pinia";
 import { computed } from "vue";
 
+import {
+  MAX_LENGTH_FOR_TEXT_INPUT,
+  MIN_LENGTH_FOR_BEGRUENDUNG,
+} from "@/constants.ts";
 import { useStimmabgabevermerkeStore } from "@/stores/stimmabgabevermerkeStore.ts";
 import { useUserStore } from "@/stores/userStore.ts";
 import { useWahlenStore } from "@/stores/wahlenStore.ts";
 import { useWahlscheineStore } from "@/stores/wahlscheineStore.ts";
 import { EingenommenerWahlscheinStimmzettelartEnum } from "@/types/stimmabgabevermerke/EingenommenerWahlscheinStimmzettelartEnum.ts";
 
-export function useDifferenceDialogUtils(wahlId: string) {
+export function useDifferenceDialogUtils(wahlId = "") {
   const { wahlscheine } = storeToRefs(useWahlscheineStore());
   const { stimmabgabevermerke } = storeToRefs(useStimmabgabevermerkeStore());
   const { wahlenActions } = useWahlenStore();
@@ -37,9 +43,19 @@ export function useDifferenceDialogUtils(wahlId: string) {
       anzahlWahlscheineOrStimmabgabevermerke.value !== anzahlStimmzettel.value
   );
 
+  function updateValidationStateForBegruendung(
+    differenceBegruendung: DifferenceBegruendung
+  ): void {
+    const begruendung = differenceBegruendung.begruendung;
+    differenceBegruendung.isBegruendungValid =
+      begruendung.length >= MIN_LENGTH_FOR_BEGRUENDUNG &&
+      begruendung.length <= MAX_LENGTH_FOR_TEXT_INPUT;
+  }
+
   return {
     anzahlWahlscheineOrStimmabgabevermerke,
     anzahlStimmzettel,
     isWahlscheineUnequalToStimmzettel,
+    updateValidationStateForBegruendung,
   };
 }
